requirements.txt
gradio
speechrecognition
gtts
google-generativeai
langdetect
pydub
